+++ api/Section.js
@@ -40,14 +40,31 @@ module.exports = [
     description: '',
   },
   {
+    name: 'index',
+    returns: 'Integer',
+    description: 'which number section is this, in the whole document.',
+  },
+  {
+    name: 'nextSibling',
+    returns: 'Section',
+    description: 'a section following this one, under the current parent: eg. 1920s → 1930s ',
+    aliases: ['next']
+  },
+  {
+    name: 'lastSibling',
+    returns: 'Section',
+    description: 'a section before this one, under the current parent: eg. 1930s → 1920s',
+    aliases: ['last', 'previous', 'previousSibling']
+  },
+  {
     name: 'children',
     returns: 'Array',
-    description: '',
+    description: 'any sections more specific than this one: eg. History → [PreHistory, 1920s, 1930s]',
   },
   {
     name: 'parent',
-    returns: 'Array',
-    description: '',
+    returns: 'Section',
+    description: 'the section, broader than this one: eg. 1920s → History ',
   },
   {
     name: 'toMarkdown',