+++ tests/table.test.js
@@ -31,7 +31,7 @@ test('rnli stations', t => {
   t.equal(intro.sentences().length > 0, true, 'intro-sentence-length');
 
   var key = doc.sections(1);
-  t.equal(key.depth, 1, 'key-depth');
+  t.equal(key.depth, 0, 'key-depth');
   t.equal(key.title, 'Key', 'key-title');
   t.equal(key.sentences().length, 0, 'key-no-sentences');
   t.deepEqual(key.images(), [], 'key-no-images');
@@ -40,7 +40,7 @@ test('rnli stations', t => {
   t.deepEqual(key.tables(), [], 'key-no-tables');
 
   var lifeboat = doc.sections(2);
-  t.equal(lifeboat.depth, 2, 'lifeboat-depth');
+  t.equal(lifeboat.depth, 1, 'lifeboat-depth');
   t.equal(lifeboat.templates().main[0], 'Royal National Lifeboat Institution lifeboats', 'lifeboat-main');
   t.equal(lifeboat.lists(0).length, 3, 'lifeboat-list');
   t.equal(lifeboat.sentences().length, 3, 'lifeboat-sentences');