+++ changelog.md
@@ -33,3 +33,4 @@
 ## 3.0.0
 * BIG API RE-WRITE!
 * wtf.fetch() uses promises, and native `fetch()` method (when available)
+* section depth values now start at 0